New and much nicer model formulation
